Automate pre-build unit testing.
This commit expands the CircleCI build automation configuration,
and adds Cloverage unit test coverage analysis, to the Skein
deployment pipeline.

Software-as-a-service components now include:

*   https://github.com/atomicsauce/skein
*   https://circleci.com/gh/atomicsauce/skein
*   https://codecov.io/gh/atomicsauce/skein

All three of the above tools are reporting out their respective
events to the Atomic Sauce Slack instance.

Instructions for pre-commit CLI invocation of all tools is now
documented at:

*   https://github.com/atomicsauce/skein/wiki/Before-you-commit

This should allow developers to confirm locally that things
are in good shape before commiting to the Github repository.

This commit corresponds to requirement 198 in the TargetProcess
tracking system.